requirements.txt
yt-dlp
ffmpeg-python
openai-whisper
transformers
tqdm
python-slugify